+++ sys/miscfs/fdesc/fdesc_vfsops.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: fdesc_vfsops.c,v 1.18 1995/03/08 01:34:45 cgd Exp $	*/
+/*	$NetBSD: fdesc_vfsops.c,v 1.19 1995/03/09 12:05:50 mycroft Exp $	*/
 
 /*
  * Copyright (c) 1992, 1993
@@ -69,7 +69,7 @@ fdesc_mount(mp, path, data, ndp, p)
 	struct proc *p;
 {
 	int error = 0;
-	u_long size;
+	size_t size;
 	struct fdescmount *fmp;
 	struct vnode *rvp;
